backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1
requests==2.31.0
cryptography==42.0.7
httpx==0.27.0
gspread==6.1.2
google-auth==2.30.0
